Add tests/unit crate and feature-gated test support foundation
