+++ ci/coverage-theoretical-max-receipts/coverage-theoretical-max.json
@@ -7,15 +7,15 @@
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "functions": {
-        "amount": 2096,
+        "amount": 2093,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "lines": {
-        "amount": 21977,
+        "amount": 21965,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "regions": {
-        "amount": 36551,
+        "amount": 36534,
         "theoretical_max_percent": "100.00000000000000000000000000"
       }
     },
@@ -25,15 +25,15 @@
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "functions": {
-        "amount": 2096,
+        "amount": 2093,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "lines": {
-        "amount": 21977,
+        "amount": 21965,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "regions": {
-        "amount": 36551,
+        "amount": 36534,
         "theoretical_max_percent": "100.00000000000000000000000000"
       }
     },
@@ -43,15 +43,15 @@
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "functions": {
-        "amount": 2096,
+        "amount": 2093,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "lines": {
-        "amount": 21977,
+        "amount": 21965,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "regions": {
-        "amount": 36551,
+        "amount": 36534,
         "theoretical_max_percent": "100.00000000000000000000000000"
       }
     }
@@ -61,13 +61,13 @@
       "amount": 1776
     },
     "functions": {
-      "amount": 2096
+      "amount": 2093
     },
     "lines": {
-      "amount": 21977
+      "amount": 21965
     },
     "regions": {
-      "amount": 36551
+      "amount": 36534
     }
   }
 }
