+++ ci/coverage-theoretical-max-receipts/coverage-theoretical-max.sha256
@@ -1,1 +1,1 @@
-8b27f0a3724571142b11d6b232c29341cd25080751263c25bf2bbf7c675b4f48  ci/coverage-theoretical-max-receipts/coverage-theoretical-max.json
+89669e8d38cb1ce5323adcddebd29165acc67e1d29c276bfb41d334ab6a41007  ci/coverage-theoretical-max-receipts/coverage-theoretical-max.json
