+++ ci/llvm-cov-receipts/linux/llvm-cov-report.txt
@@ -57,7 +57,7 @@ prompts.rs                                    70                 0   100.00%    
 release_notes.rs                             411                 0   100.00%          44                 0   100.00%         228                 0   100.00%          16                 0   100.00%
 runtime/mod.rs                               107                 0   100.00%          11                 0   100.00%          76                 0   100.00%           6                 0   100.00%
 state/mod.rs                                 108                 0   100.00%          36                 0   100.00%         108                 0   100.00%           0                 0         -
-test_support.rs                               60                 0   100.00%          10                 0   100.00%          43                 0   100.00%           2                 0   100.00%
+test_support.rs                               43                 0   100.00%           7                 0   100.00%          31                 0   100.00%           2                 0   100.00%
 tui/input/command.rs                          15                 0   100.00%           3                 0   100.00%           9                 0   100.00%           0                 0         -
 tui/input/confirm.rs                          32                 0   100.00%           6                 0   100.00%          18                 0   100.00%           0                 0         -
 tui/input/mod.rs                             118                 0   100.00%           3                 0   100.00%          59                 0   100.00%           0                 0         -
@@ -82,4 +82,4 @@ tui/render/modals/progress.rs                 99                 0   100.00%    
 tui/render/modals/settings_menu.rs            94                 0   100.00%           1                 0   100.00%          38                 0   100.00%           4                 0   100.00%
 update.rs                                    580                 0   100.00%          41                 0   100.00%         353                 0   100.00%          10                 0   100.00%
 ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-TOTAL                                      36551                 0   100.00%        2096                 0   100.00%       21977                 0   100.00%        1776                 0   100.00%
+TOTAL                                      36534                 0   100.00%        2093                 0   100.00%       21965                 0   100.00%        1776                 0   100.00%
