+++ ci/llvm-cov-receipts/windows/llvm-cov-report.sha256
@@ -1,1 +1,1 @@
-675d59058dc8f07cf614940fbb0be5fef833484b4e52ba948ff993d335a63a41  ci/llvm-cov-receipts/windows/llvm-cov-report.txt
+90453ac6b0227c602cd4b1502641e789e03d585c3eb01822e14dce99c20ca1c4  ci/llvm-cov-receipts/windows/llvm-cov-report.txt
