+++ src/lib.rs
@@ -45,8 +45,8 @@ pub mod prompts;
 pub mod release_notes;
 pub(crate) mod runtime;
 pub mod state;
-#[cfg(test)]
-pub(crate) mod test_support;
+#[cfg(any(test, feature = "test-support"))]
+pub mod test_support;
 pub mod tui;
 pub mod update;
