+++ tests/unit.rs
@@ -1,0 +1,5 @@
+//! Unit tests migrated out of `src/`.
+
+mod unit {
+    pub mod support;
+}
